code clean up and adding test cases

+++ src/main/java/com/pinet/app/service/UserTypeService.java
@@ -3,6 +3,7 @@ package com.pinet.app.service;
 import com.pinet.app.entities.ServiceTypeEntity;
 import com.pinet.app.entities.UserDataEntity;
 import com.pinet.app.model.UserDataResponse;
+import com.pinet.app.model.UserDataVO;
 import com.pinet.app.repository.UserDataRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,39 +16,22 @@ import java.util.List;
  * Created by ashmeet on 7/19/17.
  */
 @Service
-public class UserTypeService implements UserTypeServiceInteface {
+public class UserTypeService {
 
 
     @Autowired
     UserDataRepository userDataRepository;
 
-    @Override
-    public UserDataResponse saveUser(UserDataEntity userDataEntity){
 
-//        userDataEntity.setName(name);
-//        userDataEntity.setUsername(username);
-//        userDataEntity.setDob(new Date());
-//        userDataEntity.setGender(gender);
-//        userDataEntity.setNationality(nationality);
-//        userDataEntity.setCitizenNo(citizenNo);
-//        userDataEntity.setPassportNo(passportNo);
-//        userDataEntity.setFatherName(fatherName);
-//        userDataEntity.setEmail(email);
-//        userDataEntity.setAddress(address);
-//        userDataEntity.setMobileNo(mobileNo);
-//        userDataEntity.setPhoneNo(phoneNo);
-//        userDataEntity.setOccupancyType(occupationType);
-//        userDataEntity.setServiceCode(serviceCode);
-//        userDataEntity.setSubscriptionType(subscriptionType);
-//        userDataEntity.setInstallationAddress(installationAddress);
+    public UserDataResponse saveUser(UserDataVO userDataVO) {
+        UserDataEntity userDataEntity = new UserDataEntity(userDataVO);
 
-        UserDataEntity savedUser=userDataRepository.save(userDataEntity);
+        UserDataEntity savedUser = userDataRepository.save(userDataEntity);
         UserDataResponse response = new UserDataResponse(savedUser.getName(), savedUser.getUsername(), savedUser.getGender(), savedUser.getDob(), savedUser.getNationality(), savedUser.getCitizenNo(), savedUser.getPassportNo(), savedUser.getFatherName(), savedUser.getEmail(), savedUser.getAddress(), savedUser.getMobileNo(), savedUser.getPhoneNo(), savedUser.getOccupancyType(), savedUser.getServiceCode(), savedUser.getSubscriptionType(), savedUser.getInstallationAddress());
         return response;
     }
 
-    @Override
-    public List<UserDataResponse> getUsers() {
+    public List<UserDataResponse> getAllUsers() {
         List<UserDataEntity> userDataEntityList = userDataRepository.findAll();
         List<UserDataResponse> userList = new ArrayList<>();
 
@@ -59,44 +43,24 @@ public class UserTypeService implements UserTypeServiceInteface {
         return userList;
     }
 
-    @Override
-    public UserDataResponse getUserByUserName(String username) {
-        return null;
-    }
 
-    @Override
     public UserDataResponse getUserById(Integer userId) {
-        UserDataEntity userDataEntity=new UserDataEntity();
-
-
-        userDataEntity = userDataRepository.findOne(userId);
-
-        UserDataEntity savedUser=userDataRepository.save(userDataEntity);
-
-        UserDataResponse response = new UserDataResponse(savedUser.getUserId(), savedUser.getName(), savedUser.getUsername(), savedUser.getGender(), savedUser.getDob(), savedUser.getNationality(), savedUser.getCitizenNo(), savedUser.getPassportNo(), savedUser.getFatherName(), savedUser.getEmail(), savedUser.getAddress(), savedUser.getMobileNo(), savedUser.getPhoneNo(), savedUser.getOccupancyType(), savedUser.getServiceCode(), savedUser.getSubscriptionType(), savedUser.getInstallationAddress());
+        UserDataEntity userDataEntity = userDataRepository.findOne(userId);
+        UserDataResponse response = new UserDataResponse(userDataEntity.getUserId(), userDataEntity.getName(), userDataEntity.getUsername(), userDataEntity.getGender(), userDataEntity.getDob(), userDataEntity.getNationality(), userDataEntity.getCitizenNo(), userDataEntity.getPassportNo(), userDataEntity.getFatherName(), userDataEntity.getEmail(), userDataEntity.getAddress(), userDataEntity.getMobileNo(), userDataEntity.getPhoneNo(), userDataEntity.getOccupancyType(), userDataEntity.getServiceCode(), userDataEntity.getSubscriptionType(), userDataEntity.getInstallationAddress());
 
         return response;
     }
 
-    @Override
-    public UserDataResponse deleteUserById(Integer userId) {
-        UserDataEntity user=new UserDataEntity();
-
-
-        user = userDataRepository.findOne(userId);
-
-//        UserDataEntity savedUser=userDataRepository.save(userDataEntity);
-
-        userDataRepository.delete(user);
-
-        System.out.println("UserDeleted");
-
-        UserDataResponse response = new UserDataResponse(user.getName(), user.getUsername(), user.getGender(), user.getDob(), user.getNationality(), user.getCitizenNo(), user.getPassportNo(), user.getFatherName(), user.getEmail(), user.getAddress(), user.getMobileNo(), user.getPhoneNo(), user.getOccupancyType(), user.getServiceCode(), user.getSubscriptionType(), user.getInstallationAddress());
-//
-        return response;
+    public String deleteUserById(Integer userId) {
+        UserDataEntity user = userDataRepository.findOne(userId);
+        if (user == null) {
+            return "User not found";
+        } else {
+            userDataRepository.delete(user);
+            return "User successfully deleted";
+        }
     }
 
-    @Override
     public UserDataResponse updateUserById(UserDataEntity userDataEntity) {
         UserDataEntity user = new UserDataEntity();
         user = userDataRepository.findOne(userDataEntity.getUserId());
@@ -118,7 +82,7 @@ public class UserTypeService implements UserTypeServiceInteface {
         user.setSubscriptionType(userDataEntity.getSubscriptionType());
         user.setInstallationAddress(userDataEntity.getInstallationAddress());
 
-        UserDataEntity savedUser=userDataRepository.save(user);
+        UserDataEntity savedUser = userDataRepository.save(user);
         UserDataResponse response = new UserDataResponse(savedUser.getName(), savedUser.getUsername(), savedUser.getGender(), savedUser.getDob(), savedUser.getNationality(), savedUser.getCitizenNo(), savedUser.getPassportNo(), savedUser.getFatherName(), savedUser.getEmail(), savedUser.getAddress(), savedUser.getMobileNo(), savedUser.getPhoneNo(), savedUser.getOccupancyType(), savedUser.getServiceCode(), savedUser.getSubscriptionType(), savedUser.getInstallationAddress());
         return response;
     }
